Extract renderLogColumn to shared log-utils module
